app/requirements.txt
celery==5.3.6
cryptography==41.0.5
drf-yasg==1.21.7
django==4.2.7
django_celery_results==2.5.1
django-debug-toolbar==4.2.0
django-split-settings==1.2.0
djangorestframework==3.14.0
flake8==6.1.0
flower==2.0.1
gunicorn==21.2.0
python-dotenv==1.0.0
pymysql==1.1.0
redis==5.0.1
